SONAR-9871 move webhook code to use it in web
as well as we use it in CE report processing
