+++ server/sonar-server/src/main/java/org/sonar/server/webhook/WebhookDelivery.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation.task.projectanalysis.webhook;
+package org.sonar.server.webhook;
 
 import java.util.Optional;
 import javax.annotation.Nullable;
