+++ server/sonar-server/src/test/java/org/sonar/server/webhook/WebHooksImplTest.java
@@ -1,0 +1,120 @@
+/*
+ * SonarQube
+ * Copyright (C) 2009-2017 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonar.server.webhook;
+
+import java.io.IOException;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import org.junit.Rule;
+import org.junit.Test;
+import org.sonar.api.config.internal.MapSettings;
+import org.sonar.api.utils.log.LogTester;
+import org.sonar.api.utils.log.LoggerLevel;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
+
+public class WebHooksImplTest {
+
+  private static final long NOW = 1_500_000_000_000L;
+  private static final String PROJECT_UUID = "P1_UUID";
+
+  @Rule
+  public LogTester logTester = new LogTester();
+
+  private final MapSettings settings = new MapSettings();
+  private final TestWebhookCaller caller = new TestWebhookCaller();
+  private final WebhookDeliveryStorage deliveryStorage = mock(WebhookDeliveryStorage.class);
+  private final WebhookPayload mock = mock(WebhookPayload.class);
+  private final WebHooksImpl underTest = new WebHooksImpl(caller, deliveryStorage);
+
+  @Test
+  public void do_nothing_if_no_webhooks() {
+    underTest.sendProjectAnalysisUpdate(settings.asConfig(), new WebHooks.Analysis(PROJECT_UUID, "#1"), () -> mock);
+
+    assertThat(caller.countSent()).isEqualTo(0);
+    assertThat(logTester.logs(LoggerLevel.DEBUG)).isEmpty();
+    verifyZeroInteractions(deliveryStorage);
+  }
+
+  @Test
+  public void send_global_webhooks() {
+    settings.setProperty("sonar.webhooks.global", "1,2");
+    settings.setProperty("sonar.webhooks.global.1.name", "First");
+    settings.setProperty("sonar.webhooks.global.1.url", "http://url1");
+    settings.setProperty("sonar.webhooks.global.2.name", "Second");
+    settings.setProperty("sonar.webhooks.global.2.url", "http://url2");
+    caller.enqueueSuccess(NOW, 200, 1_234);
+    caller.enqueueFailure(NOW, new IOException("Fail to connect"));
+
+    underTest.sendProjectAnalysisUpdate(settings.asConfig(), new WebHooks.Analysis(PROJECT_UUID, "#1"), () -> mock);
+
+    assertThat(caller.countSent()).isEqualTo(2);
+    assertThat(logTester.logs(LoggerLevel.DEBUG)).contains("Sent webhook 'First' | url=http://url1 | time=1234ms | status=200");
+    assertThat(logTester.logs(LoggerLevel.DEBUG)).contains("Failed to send webhook 'Second' | url=http://url2 | message=Fail to connect");
+    verify(deliveryStorage, times(2)).persist(any(WebhookDelivery.class));
+    verify(deliveryStorage).purge(PROJECT_UUID);
+  }
+
+  @Test
+  public void send_project_webhooks() {
+    settings.setProperty("sonar.webhooks.project", "1");
+    settings.setProperty("sonar.webhooks.project.1.name", "First");
+    settings.setProperty("sonar.webhooks.project.1.url", "http://url1");
+    caller.enqueueSuccess(NOW, 200, 1_234);
+
+    underTest.sendProjectAnalysisUpdate(settings.asConfig(), new WebHooks.Analysis(PROJECT_UUID, "#1"), () -> mock);
+
+    assertThat(caller.countSent()).isEqualTo(1);
+    assertThat(logTester.logs(LoggerLevel.DEBUG)).contains("Sent webhook 'First' | url=http://url1 | time=1234ms | status=200");
+    verify(deliveryStorage).persist(any(WebhookDelivery.class));
+    verify(deliveryStorage).purge(PROJECT_UUID);
+  }
+
+  @Test
+  public void process_only_the_10_first_global_webhooks() {
+    testMaxWebhooks("sonar.webhooks.global");
+  }
+
+  @Test
+  public void process_only_the_10_first_project_webhooks() {
+    testMaxWebhooks("sonar.webhooks.project");
+  }
+
+  private void testMaxWebhooks(String property) {
+    IntStream.range(1, 15)
+      .forEach(i -> {
+        settings.setProperty(property + "." + i + ".name", "First");
+        settings.setProperty(property + "." + i + ".url", "http://url");
+        caller.enqueueSuccess(NOW, 200, 1_234);
+      });
+    settings.setProperty(property, IntStream.range(1, 15).mapToObj(String::valueOf).collect(Collectors.joining(",")));
+
+    underTest.sendProjectAnalysisUpdate(settings.asConfig(), new WebHooks.Analysis(PROJECT_UUID, "#1"), () -> mock);
+
+    assertThat(caller.countSent()).isEqualTo(10);
+    assertThat(logTester.logs(LoggerLevel.DEBUG).stream().filter(log -> log.contains("Sent"))).hasSize(10);
+  }
+
+}
