+++ server/sonar-server/src/test/java/org/sonar/server/webhook/WebhookCallerImplTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation.task.projectanalysis.webhook;
+package org.sonar.server.webhook;
 
 import okhttp3.HttpUrl;
 import okhttp3.mockwebserver.MockResponse;
