+++ server/sonar-server/src/test/java/org/sonar/server/webhook/WebhookDeliveryStorageTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation.task.projectanalysis.webhook;
+package org.sonar.server.webhook;
 
 import java.io.IOException;
 import org.junit.Rule;
@@ -28,6 +28,10 @@ import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
 import org.sonar.db.DbTester;
 import org.sonar.db.webhook.WebhookDeliveryDto;
+import org.sonar.server.webhook.Webhook;
+import org.sonar.server.webhook.WebhookDelivery;
+import org.sonar.server.webhook.WebhookDeliveryStorage;
+import org.sonar.server.webhook.WebhookPayload;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
