+++ server/sonar-server/src/test/java/org/sonar/server/webhook/WebhookModuleTest.java
@@ -17,16 +17,29 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation.task.projectanalysis.webhook;
+package org.sonar.server.webhook;
 
-import org.sonar.api.ce.posttask.PostProjectAnalysisTask;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+import org.sonar.core.platform.ComponentContainer;
 
-public class TestWebhookPayloadFactory implements WebhookPayloadFactory {
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.sonar.core.platform.ComponentContainer.COMPONENTS_IN_EMPTY_COMPONENT_CONTAINER;
 
-  private static final String FAKE_JSON = "{\"payload\": true}";
+public class WebhookModuleTest {
 
-  @Override
-  public WebhookPayload create(PostProjectAnalysisTask.ProjectAnalysis analysis) {
-    return new WebhookPayload(analysis.getProject().getKey(), FAKE_JSON);
+  @Rule
+  public ExpectedException expectedException = ExpectedException.none();
+
+  private WebhookModule underTest = new WebhookModule();
+
+  @Test
+  public void verify_count_of_added_components() {
+    ComponentContainer container = new ComponentContainer();
+
+    underTest.configure(container);
+
+    assertThat(container.size()).isEqualTo(3 + COMPONENTS_IN_EMPTY_COMPONENT_CONTAINER);
   }
 }
